fetch is now available as a polyfill or a self-contained library
