+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 bower_components/
+library/
 node_modules/
