+++ package.json
@@ -13,12 +13,14 @@
     "bower": "1.3.8",
     "chai": "1.10.0",
     "jshint": "2.5.2",
-    "mocha-phantomjs": "3.5.2",
     "mocha": "2.1.0",
-    "phantomjs": "1.9.13"
+    "mocha-phantomjs": "3.5.2",
+    "phantomjs": "1.9.13",
+    "webpack": "1.11.0"
   },
-  "files" : [
+  "files": [
     "LICENSE",
+    "library/",
     "fetch.js"
   ]
 }
